serial_server/requirements.txt
pyserial-asyncio
aiohttp